Remove Top nav or hidden In admin page And A Sidebar Trigger button
