+++ src/app/admin/layout.tsx
@@ -1,92 +1,30 @@
 
-'use client';
-
-import { useAuth } from '@/hooks/useAuth';
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-import { Loader2, LayoutDashboard, Package, ShoppingCart, Image as ImageIcon, Users } from 'lucide-react';
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-import { cn } from '@/lib/utils';
-import { Button } from '@/components/ui/button';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/ui/tooltip';
-import AdminHeader from '@/components/layout/AdminHeader';
+import type { Metadata } from 'next';
+import { TooltipProvider } from '@/components/ui/tooltip';
 import { Toaster } from '@/components/ui/toaster';
+import AdminPanel from './AdminPanel';
+import { AuthProvider } from '@/hooks/useAuth'; // Required for auth checks in AdminPanel
 
-
-const navItems = [
-  { href: '/admin', label: 'Dashboard', icon: LayoutDashboard },
-  { href: '/admin/products', label: 'Products', icon: Package },
-  { href: '/admin/orders', label: 'Orders', icon: ShoppingCart },
-  { href: '/admin/users', label: 'Users', icon: Users },
-  { href: '/admin/banners', label: 'Banners', icon: ImageIcon },
-];
+export const metadata: Metadata = {
+  title: 'Evanie Glow Admin',
+  description: 'Admin dashboard for Evanie Glow.',
+};
 
 export default function AdminLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const { user, loading } = useAuth();
-  const router = useRouter();
-  const pathname = usePathname();
-
-  useEffect(() => {
-    if (!loading) {
-      if (!user) {
-        router.push('/login');
-      } else if (user.role !== 'ADMIN') {
-        router.push('/account'); // Redirect non-admins
-      }
-    }
-  }, [user, loading, router]);
-
-  if (loading || !user || user.role !== 'ADMIN') {
-    return (
-      <div className="flex h-screen items-center justify-center bg-background">
-        <Loader2 className="h-8 w-8 animate-spin" />
-      </div>
-    );
-  }
-
   return (
-    <TooltipProvider>
-      <div className="flex min-h-screen w-full flex-col bg-muted/40">
-        <aside className="fixed inset-y-0 left-0 z-10 hidden w-14 flex-col border-r bg-background sm:flex">
-          <nav className="flex flex-col items-center gap-4 px-2 sm:py-5">
-            <Link href="/admin" className="group flex h-9 w-9 shrink-0 items-center justify-center gap-2 rounded-full bg-primary text-lg font-semibold text-primary-foreground md:h-8 md:w-8 md:text-base">
-              <span className="font-bold font-headline text-lg">G</span>
-              <span className="sr-only">Evanie Glow Admin</span>
-            </Link>
-            {navItems.map((item) => (
-              <Tooltip key={item.href}>
-                <TooltipTrigger asChild>
-                  <Link href={item.href} className={cn(
-                    "flex h-9 w-9 items-center justify-center rounded-lg text-muted-foreground transition-colors hover:text-foreground md:h-8 md:w-8",
-                    pathname === item.href && "bg-accent text-accent-foreground"
-                  )}>
-                    <item.icon className="h-5 w-5" />
-                    <span className="sr-only">{item.label}</span>
-                  </Link>
-                </TooltipTrigger>
-                <TooltipContent side="right">{item.label}</TooltipContent>
-              </Tooltip>
-            ))}
-          </nav>
-        </aside>
-        <div className="flex flex-col sm:gap-4 sm:py-4 sm:pl-14">
-          <AdminHeader />
-          <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8">
-            {children}
-          </main>
-        </div>
-      </div>
-      <Toaster />
-    </TooltipProvider>
+    <html lang="en" suppressHydrationWarning>
+      <body className="font-sans antialiased">
+         <AuthProvider>
+            <TooltipProvider>
+              <AdminPanel>{children}</AdminPanel>
+              <Toaster />
+            </TooltipProvider>
+         </AuthProvider>
+      </body>
+    </html>
   );
 }
